Resolve LEL-014

+++ core/data/src/main/java/io/github/faening/lello/core/data/repository/JournalCategoryRepository.kt
@@ -1,7 +1,7 @@
 package io.github.faening.lello.core.data.repository
 
 import io.github.faening.lello.core.database.dao.JournalCategoryDao
-import io.github.faening.lello.core.database.model.toModel
+import io.github.faening.lello.core.database.model.option.toModel
 import io.github.faening.lello.core.domain.repository.JournalCategoryResources
 import io.github.faening.lello.core.model.journal.JournalCategory
 import kotlinx.coroutines.flow.Flow
@@ -28,7 +28,7 @@ class JournalCategoryRepository @Inject constructor(
             .map { list -> list.map { it.toModel() } }
     }
 
-    override fun getById(id: Int): Flow<JournalCategory>? {
+    override fun getById(id: Long): Flow<JournalCategory>? {
         return journalCategoryDao.getById(id)?.map { it.toModel() }
     }
 }
